feat: Implement the InstructionMenu component for the intructor dashboard
